Convert report comment to a dropdown item
